� A Deeper Look at v-model in Custom Components When I use v-model with a custom component, it creates two-way communication between the parent and the child component.
 1. Receiving the value from the parent (binding)
In the child component, I need to define a prop called modelValue:
props: {
  modelValue: String
}
This modelValue will automatically receive the value passed from the parent like this:
<CustomComponent v-model='currentTag' />
So modelValue becomes equal to currentTag, and I can use it in the child however I need.

 2. Telling the parent when the value changes (emitting)
When the value changes in the child (like on button click), I need to notify the parent by emitting this:
@click='('update:modelValue', tag)'
This will update the parents v-model value automatically.

 3. Easier & cleaner way: use custom name with v-model
Instead of using modelValue, Vue allows me to give it a custom name:
<CustomComponent v-model:currentTag='currentTag' />
props: {
  currentTag: String
}
And emit like this:
@click='('update:currentTag', tag)' and use currentTag rather then modelValue in the component